custom events

+++ api/src/app/controllers/ContactController.js
@@ -12,7 +12,7 @@ class ContactController {
     const { id } = request.params;
 
     if (!isValidUUID(id)) {
-      return response.status(400).json({ error: 'Invalid user id' })
+      return response.status(400).json({ error: 'Invalid contact id' })
     }
 
     const contact = await ContactsRepository.findById(id);
@@ -55,7 +55,7 @@ class ContactController {
     const { name, email, phone, category_id } = request.body;
 
     if (!isValidUUID(id)) {
-      return response.status(400).json({ error: 'Invalid user id' })
+      return response.status(400).json({ error: 'Invalid contact id' })
     }
 
     if (category_id && !isValidUUID(category_id)) {
@@ -64,6 +64,14 @@ class ContactController {
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
+    }
+
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+
+      if (contactExists) {
+        return response.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contactById = await ContactsRepository.findById(id);
@@ -79,13 +87,17 @@ class ContactController {
     }
 
     const contact = await ContactsRepository.update(id, {
-      name, email, phone, category_id: category_id || null,
+      name, email: email || null, phone, category_id: category_id || null,
     });
     response.json(contact);
   }
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' })
+    }
 
     await ContactsRepository.delete(id);
     response.sendStatus(204);
